[Cryptle] Express traded pair as (asset, currency)
Original representation of pair was directly coupled with the Bitstamp
Websocket API representation. This would make it very hard to introduce
new exchanges. Even interfacing with the Bitstamp REST API proved to be
a problem.

The new representation decouples exchange representation of traded pairs
by introducing an internal representation in Cryptle. Portfolio class
is heavily affected as well.