Partial fix of whenexec
The intended behaviour of registry control flow is not achieved in this fix. A major overhaul of the
data type returned by aggregator (or equivalent) needs to be reviewed in order to allow retrieval of the real close time.

Actual updating at close probably is physically impossible (due to the nature of "close price" is
unknown before the bar elapses).