pivot.py adopted to new Registry and executable
This commit is not a functional one. Although buy and sell activities are present, they are not
conforming to the referencePivot.log.

Moreover, it is observed that the running speed of backtests is significantly slower after adding
transtitions dependency. More investigation might be warranted to inspect the impact of using
transitions. An inhouse framework might be more appropriate in our use case given our highly
restricted need of a StateMachine.