Implement command line interface for Strategy checking (#15)
First PR for  building command line interface for Strategy checking

There would be more in the future for validating future Strategies developed, and also to keep up with the Cryptle development. It should be noted that the CLI Click tests should also be maintained and updated whenever there are changes to the core Cryptle facilities.

The CLI of Cryptle primarily uses Click. A new directory cryptle/test is created for hosting the relevant files.

Apart from the basic structure of the test suite, several basic checks were implemented for demonstration. Their behaviour are visually validated:
-   Check whether the Strategy base __init__ method was called
-   Check whether there are same number of aggregators and candles
    declared within the Strategy
    -   Could also consider checking whether they have same
        period/lookback value in the future
- Add checking to function signature of codeblocks

Miscellaneous:
* Filedir argument accepts both a.b and path syntax
* Make cryptle.test a subpackage of Cryptle